refactor(cli): split supply chain conformance command
